api/requirements/development.txt
django-extensions>=3.2.3,<3.3
django-debug-toolbar>=4.3.0,<4.4